refactor(vscode): isolate diagnostic commands (#4255) (#4389)
